Only migrate processes
