+++ source/databricks/tests/datamigration/migration_scripts/test_202212051100_Move_Results_To_Wholesale_Container.py
@@ -20,7 +20,7 @@ from package.datamigration.migration_script_args import MigrationScriptArgs
 
 
 @patch(
-    "package.datamigration.migration_scripts.202211281100_Move_To_Wholesale_Container.DataLakeDirectoryClient"
+    "package.datamigration.migration_scripts.202212051100_Move_Results_To_Wholesale_Container.DataLakeDirectoryClient"
 )
 def test__apply__directory_client_contructed_with_correct_arguments(
     mock_directory_client,
@@ -30,28 +30,19 @@ def test__apply__directory_client_contructed_with_correct_arguments(
     sut = get_migration_script()
     migration_args = MigrationScriptArgs("", "", None)
     processes_container = "processes"
-    events_container = "integration-events"
     results_source_directory = "results"
-    events_source_directory = "events"
-    events_checkpoint_source_directory = "events-checkpoint"
-
-    result_call = call(ANY, processes_container, results_source_directory, ANY)
-    event_call = call(ANY, events_container, events_source_directory, ANY)
-    events_checkpoint_call = call(
-        ANY, events_container, events_checkpoint_source_directory, ANY
-    )
-
-    expected_calls = [result_call, event_call, events_checkpoint_call]
 
     # Act
     sut.apply(migration_args)
 
     # Assert
-    mock_directory_client.assert_has_calls(expected_calls, any_order=True)
+    mock_directory_client.assert_called_with(
+        ANY, processes_container, results_source_directory, ANY
+    )
 
 
 @patch(
-    "package.datamigration.migration_scripts.202211281100_Move_To_Wholesale_Container.DataLakeDirectoryClient"
+    "package.datamigration.migration_scripts.202212051100_Move_Results_To_Wholesale_Container.DataLakeDirectoryClient"
 )
 def test__apply__calls_rename_directory_with_correct_arguments(
     mock_directory_client,
@@ -61,21 +52,18 @@ def test__apply__calls_rename_directory_with_correct_arguments(
     sut = get_migration_script()
     migration_args = MigrationScriptArgs("", "", None)
     mock_directory_client.return_value.exists.return_value = True
-    expected_calls = [
-        call(new_name="wholesale/results"),
-        call(new_name="wholesale/events"),
-        call(new_name="wholesale/events-checkpoint"),
-    ]
 
     # Act
     sut.apply(migration_args)
 
     # Assert
-    mock_directory_client.return_value.rename_directory.assert_has_calls(expected_calls)
+    mock_directory_client.return_value.rename_directory.assert_called_with(
+        new_name="wholesale/results"
+    )
 
 
 @patch(
-    "package.datamigration.migration_scripts.202211281100_Move_To_Wholesale_Container.DataLakeDirectoryClient"
+    "package.datamigration.migration_scripts.202212051100_Move_Results_To_Wholesale_Container.DataLakeDirectoryClient"
 )
 def test__apply__when_source_directory_not_exist__never_call_rename_directory(
     mock_directory_client,
@@ -95,5 +83,5 @@ def test__apply__when_source_directory_not_exist__never_call_rename_directory(
 
 def get_migration_script():
     return importlib.import_module(
-        "package.datamigration.migration_scripts.202211281100_Move_To_Wholesale_Container"
+        "package.datamigration.migration_scripts.202212051100_Move_Results_To_Wholesale_Container"
     )
